chore(editor-state): move files to viewport and history features
