+++ packages/editor/packages/editor-state/src/features/viewport/centerViewportOnCodeBlock.test.ts
@@ -1,7 +1,8 @@
 import { describe, it, expect } from 'vitest';
 
 import centerViewportOnCodeBlock, { CodeBlockBounds } from './centerViewportOnCodeBlock';
-import { createMockViewport } from './testingUtils/testUtils';
+
+import { createMockViewport } from '../../pureHelpers/testingUtils/testUtils';
 
 /**
  * Helper function to create a mock code block with essential properties for viewport centering tests
